Cleanup template; Add convert; Add icon; Add balance list;

+++ app/src/main/java/data/remote/dto/RatesDto.kt
@@ -47,7 +47,7 @@ data class RatesDto(
     val EGP: Double,
     val ERN: Double,
     val ETB: Double,
-    val EUR: Int,
+    val EUR: Double,
     val FJD: Double,
     val FKP: Double,
     val GBP: Double,
